fix: unify value stories and fix shadow issues with auto scaling
